==> 04-Apply-med7-on-Clinical-Notes.ipynb <==
from google.colab import drive
drive.mount("/content/drive", force_remount=True)

import sys
sys.path.append('/content/drive/MyDrive/Colab Notebooks/CS598_DLH_Paper211')

PATH = '/content/drive/MyDrive/Colab Notebooks/CS598_DLH_Paper211'

#! pip install https://huggingface.co/kormilitzin/en_core_med7_lg/resolve/main/en_core_med7_lg-any-py3-none-any.whl

import pandas as pd
import spacy

med7 = spacy.load("en_core_med7_lg")

preprocessed_df = pd.read_pickle(PATH + "/data/preprocessed_notes.p")

print(preprocessed_df.shape)
print(len(set(preprocessed_df.SUBJECT_ID.to_list())))

preprocessed_df['ner'] = None

count = 0
preprocessed_index = {}
for i in preprocessed_df.itertuples():

    if count % 1000 == 0:
        print(count)

    count += 1
    ind = i.Index
    text = i.preprocessed_text

    all_pred = []
    for each_sent in text:
        try:
            doc = med7(each_sent)
            result = ([(ent.text, ent.label_) for ent in doc.ents])
            if len(result) == 0:
                continue
            all_pred.append(result)
        except:
            print("error..")
            continue
    preprocessed_df.at[ind, 'ner'] = all_pred


pd.to_pickle(preprocessed_df, PATH + "/data/ner_df.p")

preprocessed_df.shape

preprocessed_df.columns


preprocessed_df

med7.component_names
