==> 10-GetModelResults.ipynb <==
import glob
from pathlib import Path
import pandas as pd
import os
import numpy as np
import seaborn as sns
from matplotlib import pyplot as plt
pd.set_option('precision', 2)


from google.colab import drive

try:
  drive.mount("/content/drive")
except:
  pass

import sys

BASE_PATH = '/content/drive/MyDrive/Colab Notebooks/CS598_DLH_Paper211'
sys.path.append(BASE_PATH)

def get_mean_stats(df):
  # df_groupby = df.groupby(['experiment','target_problem','embedding']).agg({'auc':['mean'],'auprc':['mean'], 'F1':['mean']})
  if 'Experiment' in df.columns:
    df_groupby = df.groupby(['Experiment','Task','Embedding']).agg({'auc':['mean'],'auprc':['mean'], 'F1':['mean']}).reset_index()
  else:
    df_groupby = df.groupby(['Task','Embedding']).agg({'auc':['mean'],'auprc':['mean'], 'F1':['mean']}).reset_index()

  return df_groupby

def combine_col_names(df):
  new_columns = []
  for i in df.columns:
    j=''
    if i[1]:
      j='_'.join(list(i))
    else:
      j=i[0]
    new_columns.append(j)
  df.columns = new_columns

  df['auc_mean'] = df['auc_mean'].multiply(100).round(2)
  df['auprc_mean'] = df['auprc_mean'].multiply(100).round(2)
  df['F1_mean'] = df['F1_mean'].multiply(100).round(2)

  return df

def get_results(folder_path, embedding = True):
  results_dict = {'Task':[],'Embedding':[], 'F1':[],'acc':[],'auc':[],'auprc':[]}
  files = glob.glob(f"{folder_path}/*.p")
  for file in files:
    file_name = Path(file).stem
    file_name_items = file_name.split('-')

    if embedding:
      results_dict['Task'].append(file_name_items[3])
      results_dict['Embedding'].append(file_name_items[2])
    else:
      results_dict['Task'].append(file_name_items[2])
      
    scores_dict = pd.read_pickle(file)
    results_dict['acc'].append(scores_dict['acc'])
    results_dict['auc'].append(scores_dict['auc'])
    results_dict['auprc'].append(scores_dict['auprc'])
    results_dict['F1'].append(scores_dict['F1'])
  
  df=pd.DataFrame.from_dict(results_dict,orient='index').transpose()
  print(df.columns)
  df = get_mean_stats(df)
  df = combine_col_names(df)

  return df

def get_results_subfolder(folder_path):
  subfolders = [ f.path for f in os.scandir(folder_path) if f.is_dir() ]
  results_dict = {'Experiment':[],'Embedding':[],'Task':[],'F1':[],'acc':[],'auc':[],'auprc':[]}
  for path in subfolders:
    folder_name = path.split('/')[-1]
    exclude_list = ['.ipynb_checkpoints']
    if folder_name not in exclude_list:
      files = glob.glob(f"{path}/*.p")
      for file in files:
        file_name = Path(file).stem
        file_name_items = file_name.split('-')
        results_dict['Experiment'].append(file_name_items[-1])
        results_dict['Embedding'].append(file_name_items[2])
        results_dict['Task'].append(file_name_items[3])
        scores_dict = pd.read_pickle(file)
        results_dict['acc'].append(scores_dict['acc'])
        results_dict['auc'].append(scores_dict['auc'])
        results_dict['auprc'].append(scores_dict['auprc'])
        results_dict['F1'].append(scores_dict['F1'])
  df=pd.DataFrame.from_dict(results_dict,orient='index').transpose()
  df = get_mean_stats(df)
  df = combine_col_names(df)
  return df

folder_path = '/content/drive/MyDrive/Colab Notebooks/CS598_DLH_Paper211/results/proposed/'
df_proposed = get_results_subfolder(folder_path)


df_proposed.columns = ['Experiment', 'Task', 'Embedding', 'AUROC', 'AUPRC', 'F1']
set(df_proposed.Experiment.values.tolist())



df_proposed_normal = df_proposed.loc[df_proposed['Experiment'] == 'normal']
df_proposed_sqd_adagrad = df_proposed.loc[df_proposed['Experiment'].isin(['sgd','adagrad'])]
df_proposed_100epochs = df_proposed.loc[df_proposed['Experiment'] == '100epochs']
df_proposed_100epochsstride1kernel2 = df_proposed.loc[df_proposed['Experiment'] == '100epochsstride1kernel2']
df_proposed_245 = df_proposed.loc[df_proposed['Experiment'] == '245']
df_proposed_100epochs_do_05 = df_proposed.loc[df_proposed['Experiment'] == 'test']


# df_normal=df_normal.loc[:, df_normal.columns != 'Experiment']
# df_100epochs=df_100epochs.loc[:, df_100epochs.columns != 'Experiment']
# df_245=df_245.loc[:, df_245.columns != 'Experiment']


def highlight_max(s):
    if s.dtype == np.object:
        is_max = [False for _ in range(s.shape[0])]
    else:
        is_max = s == s.max()
    return ['background: lightgreen' if cell else '' for cell in is_max]

# df_normal.style.apply(highlight_max)

df_proposed_normal[['Task',	'Embedding',	'AUROC',	'AUPRC',	'F1']].style.hide_index()

df_proposed_normal.groupby(['Experiment','Task'])[['auc_mean','auprc_mean','F1_mean']].max()

df_proposed_100epochs.groupby(['Experiment','Task'])[['auc_mean','auprc_mean','F1_mean']].max()

df_proposed_100epochsstride1kernel2.groupby(['Experiment','Task'])[['auc_mean','auprc_mean','F1_mean']].max()

df_proposed_sqd_adagrad.groupby(['Task','Experiment'])[['auc_mean','auprc_mean','F1_mean']].max()

df_proposed_100epochsstride1kernel2.groupby(['Experiment','Task'])[['auc_mean','auprc_mean','F1_mean']].max()

df_proposed_100epochs.groupby(['Experiment','Task'])[['auc_mean','auprc_mean','F1_mean']].max()

df_proposed_normal.groupby(['Experiment','Task'])[['auc_mean','auprc_mean','F1_mean']].max().reset_index()

df_proposed_245 = df_proposed_245[df_proposed_245.Task.isin(['los_2','los_4','los_5'])]
df_proposed_245

df_proposed_normal.groupby(['Experiment','Task'])[['auc_mean','auprc_mean','F1_mean']].max().reset_index()

df_proposed_100epochs_do_05.groupby(['Experiment','Task'])[['auc_mean','auprc_mean','F1_mean']].max().reset_index()

folder_path = '/content/drive/MyDrive/Colab Notebooks/CS598_DLH_Paper211/results/TimeSeriesBaseline/normal'
df_timeseries_normal = get_results(folder_path, embedding=False)

df_timeseries_normal.groupby(['Task'])[['auc_mean','auprc_mean','F1_mean']].max()
